+++ CHANGELOG.md
@@ -2,10 +2,18 @@ Change log for PC Engines sortbootorder
 =======================================
 
 Releases 4.0.x are based on PC Engines 20160304 release.
+
 Releases 4.5.x are based on mainline support submitted in
 [this gerrit ref](https://review.coreboot.org/#/c/14138/).
 
 ## [Unreleased]
+## [v4.0.4] - 2017-02-28
+### Added
+- serial console redirection option (`SgaBIOS` enable) - by default enabled
+
+### Fixed
+- fixed writing bootorder files with sizes bigger than 255 bytes
+
 ## [v4.0.3] - 2017-01-03
 ### Added
 - UART C and D toggling
@@ -45,7 +53,8 @@ initial commit based on [coreboot_140908](http://pcengines.ch/tmp/coreboot_14090
 ### Fixed
 - used proper way to access extended SPI registers
 
-[Unreleased]: https://github.com/pcengines/sortbootorder/compare/v4.0.3...coreboot-4.0.x
+[Unreleased]: https://github.com/pcengines/sortbootorder/compare/v4.0.4...coreboot-4.0.x
+[v4.0.4]: https://github.com/pcengines/sortbootorder/compare/v4.0.3...v4.0.4
 [v4.0.3]: https://github.com/pcengines/sortbootorder/compare/v4.0.2...v4.0.3
 [v4.0.2]: https://github.com/pcengines/sortbootorder/compare/v4.0.1...v4.0.2
 [v4.5.2]: https://github.com/pcengines/sortbootorder/compare/v4.0.1...v4.5.2